ai/requirements.txt
openai
python-dotenv
faiss-cpu
bs4
requests
psycopg2-binary
fastapi
uvicorn
pydantic